# perf(bitmap): change the buffer unit from `u8` to `usize`