--- a/src/expr/src/expr/expr_is_null.rs
+++ b/src/expr/src/expr/expr_is_null.rs
@@ -59,10 +59,7 @@ impl Expression for IsNullExpression {
 
     fn eval(&self, input: &DataChunk) -> Result<ArrayRef> {
         let child_arr = self.child.eval_checked(input)?;
-        let arr = BoolArray::new(
-            Bitmap::all_high_bits(input.capacity()),
-            !child_arr.null_bitmap(),
-        );
+        let arr = BoolArray::new(Bitmap::ones(input.capacity()), !child_arr.null_bitmap());
 
         Ok(Arc::new(ArrayImpl::Bool(arr)))
     }
@@ -85,7 +82,7 @@ impl Expression for IsNotNullExpression {
             Ok(child_arr) => child_arr.into_null_bitmap(),
             Err(child_arr) => child_arr.null_bitmap().clone(),
         };
-        let arr = BoolArray::new(Bitmap::all_high_bits(input.capacity()), null_bitmap);
+        let arr = BoolArray::new(Bitmap::ones(input.capacity()), null_bitmap);
 
         Ok(Arc::new(ArrayImpl::Bool(arr)))
     }