--- a/src/meta/src/stream/scale.rs
+++ b/src/meta/src/stream/scale.rs
@@ -167,10 +167,7 @@ pub(crate) fn rebalance_actor_vnode(
 
     let order_by_bitmap_desc =
         |(_, bitmap_a): &(ActorId, Bitmap), (_, bitmap_b): &(ActorId, Bitmap)| -> Ordering {
-            bitmap_a
-                .num_high_bits()
-                .cmp(&bitmap_b.num_high_bits())
-                .reverse()
+            bitmap_a.count_ones().cmp(&bitmap_b.count_ones()).reverse()
         };
 
     let builder_from_bitmap = |bitmap: &Bitmap| -> BitmapBuilder {
@@ -184,8 +181,8 @@ pub(crate) fn rebalance_actor_vnode(
     let prev_remain = removed
         .iter()
         .map(|(_, bitmap)| {
-            assert!(bitmap.num_high_bits() >= prev_expected);
-            bitmap.num_high_bits() - prev_expected
+            assert!(bitmap.count_ones() >= prev_expected);
+            bitmap.count_ones() - prev_expected
         })
         .sum::<usize>();
 
@@ -194,7 +191,7 @@ pub(crate) fn rebalance_actor_vnode(
 
     let removed_balances = removed.into_iter().map(|(actor_id, bitmap)| Balance {
         actor_id,
-        balance: bitmap.num_high_bits() as i32,
+        balance: bitmap.count_ones() as i32,
         builder: builder_from_bitmap(&bitmap),
     });
 
@@ -202,7 +199,7 @@ pub(crate) fn rebalance_actor_vnode(
         .into_iter()
         .map(|(actor_id, bitmap)| Balance {
             actor_id,
-            balance: bitmap.num_high_bits() as i32 - expected as i32,
+            balance: bitmap.count_ones() as i32 - expected as i32,
             builder: builder_from_bitmap(&bitmap),
         })
         .collect_vec();