I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

This PR changes the underlying buffer of `Bitmap` from `Vec<u8>` to `Vec<usize>`.
Processing in the unit of machine word rather than bytes makes it more efficient.

The bench results in bitmap of size 1024:

| bench | `Vec<u8>` | `Vec<usize>` | change   |
| ----- | --------- | ------------ | -------- |
| new_zeros | 22.748 ns | 18.920 ns    | -17.121% |
| new_ones  | 24.573 ns | 22.103 ns    | -10.967% |
| get   | 344.78 ps | 313.85 ps    | -9.2259% |
| and   | 105.73 ns | 26.811 ns    | -74.634% |
| or    | 105.75 ns | 26.709 ns    | -74.742% |
| not   | 30.109 ns | 23.285 ns    | -22.672% |

This PR also changes some function names and adds some docs.

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Refer to a related PR or issue link (optional)
#6868